Add the Pyenv package
We have added the necessary steps to install the Pyenv package. We have
also added a `setup_python.sh` script to install the necessary Python
version and the packages listed in the `requirements-sys.txt` file.
This is now documented in the README file.

Fish configuration files have been modified to include the necessary
variables and commands for Pyenv to work properly.