Remove `dotbot` support and Python development dependencies
Since the idea of this branch is to use Nix to manage the dotfiles as
well as the development environment, we don't need `dotbot` or the
Python development dependencies anymore.

We have only kept `dotbot` and Python independent content in the
README file.

These removals are now documented in the CHANGELOG.