Reorganize the structure of the configurations related files
Now the configurations are defined in a separate file, `system-configs.nix`,
which is imported in the `flake.nix` file.

In addition, the configurations files are now organized in separate
directories, `darwin-configurations`, `nixos-configurations` and
`home-configurations`.

With this structure, the usual way to define the configurations for the
flake are too verbose. The plan is to define a custom library to make
this process easier and cleaner.