Add an initial version of the 'configurations-manager' library
We've added an initial version of the 'configurations-manager' library,
a set of functions and utilities to easily define the configurations
handled by this repository by taking advantage of the file structure.

Note that the library structure and its the functionality should be
improved and tested before its usage. In fact, the library is not
currently used by any other part of the repository for the moment.