Restructure and document the 'configurations-manager' library
The library has been divided into multiple file to make it easier to
maintain and understand.

In addition, the functions in the `system-makers.nix` and the
`utils.nix` files have been documented. The final functions that the
library provides (in the `default.nix` file) are not yet documented.

We have also added the `home-makers.nix` file, which defines the
`mkHomeConfigurations` function, which is not yet implemented.