Add a new `usersInfo` function for home-manager extra special arguments
This function is passed to all the imports (or modules in the case of
standalone Home-Manager configurations) and when provided with the
`config` attribute set, it returns the `userInfo` attribute of the
user. With this, we can use user-specific information in the same
file.

For this to work properly in the standalone case, we've also had to add
a module that defines the `home.username` option.